Handle upgrade from DRUPAL-6--1.
The earlier version of this module (ie, the current one downloadable from Drupal.org and in use by most Drupal sites using Kaltura) does not handle the DB upgrade from that version to this.

Sites on DRUPAL-6--1 modules which upgrade without applying the .patch supplied by @webchick on http://drupal.org/node/358937 will be left with a smallint column for nid/vid. This breaks for sites with node IDs over 65535.

Closes #7.